% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{create_manifest}
\alias{create_manifest}
\title{Write a manifest file}
\usage{
create_manifest(series_dir, date = Sys.Date())
}
\arguments{
\item{series_dir}{Path to the series directory}

\item{date}{Publication date for the series}
}
\value{
A tibble
}
\description{
Write a manifest file
}
