% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image-resize.R
\name{create_resized_images}
\alias{create_resized_images}
\title{Create resized copies of all images in a directory}
\usage{
create_resized_images(
  series,
  input_folder,
  output_folder,
  pixels_wide,
  pixels_high = pixels_wide
)
}
\arguments{
\item{series}{Path to the series directory description}

\item{input_folder}{Folder within the series containing original images}

\item{output_folder}{Folder within the series to contain resized images}

\item{pixels_wide}{Width of output images in pixels}

\item{pixels_high}{Height of output images in pixels description}
}
\value{
Invisibly returns NULL. This may change
}
\description{
Create resized copies of all images in a directory
}
