% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-matrix.R
\name{generate_matrix}
\alias{generate_matrix}
\title{Generate a forces matrix}
\usage{
generate_matrix(exp, size = 80, scale = FALSE)
}
\arguments{
\item{exp}{Mat expression.}

\item{size}{Number of rows/columns to create.}

\item{scale}{Whether to apply scaling.}
}
\value{
A size x size matrix, with values equal to evaluated expression
}
\description{
Create a forces matrix based on mathematical expressions.
Taking x and y, evaluating force z.
}
\examples{
generate_matrix(2 * sin(x)^2 + sin(y)^2, 400)
}
