% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{manifest_create}
\alias{manifest_create}
\title{Write a manifest file}
\usage{
manifest_create(series, origin = bucket_local_path(), date = Sys.Date())
}
\arguments{
\item{series}{Path to the series directory}

\item{origin}{Location in which to find the series}

\item{date}{Publication date for the series}
}
\value{
A tibble
}
\description{
Write a manifest file
}
