% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{manifest_read}
\alias{manifest_read}
\title{Read the series manifest}
\usage{
manifest_read(series, origin = bucket_remote_path())
}
\arguments{
\item{series}{Name of the series}

\item{origin}{Location in which to find the series}
}
\value{
Tibble containing the manifest data
}
\description{
Read the series manifest
}
