% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{manifest_write}
\alias{manifest_write}
\title{Write a series manifest file}
\usage{
manifest_write(
  series,
  date = Sys.Date(),
  origin = bucket_local_path(),
  destination = bucket_local_path(),
  file = "manifest.csv"
)
}
\arguments{
\item{series}{Path to the series directory}

\item{date}{Publication date for the series}

\item{origin}{Location in which to find the series}

\item{destination}{Location into which the manifest is written}

\item{file}{File name for the manifest}
}
\value{
A tibble
}
\description{
Write a series manifest file
}
