% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{read_manifest}
\alias{read_manifest}
\title{Import the series manifest}
\usage{
read_manifest(series, base_url = NULL)
}
\arguments{
\item{series}{Name of the series}

\item{base_url}{URL for the bucket containing all art series}
}
\value{
Tibble containing the manifest data
}
\description{
Import the series manifest
}
