% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{repo_check}
\alias{repo_check}
\title{Check the structure of an art repository}
\usage{
repo_check(series, origin = repo_local_path())
}
\arguments{
\item{series}{Name of the series}

\item{origin}{Location in which to find the series}
}
\value{
Invisibly returns TRUE if all checks pass, FALSE if at least one
check fails
}
\description{
Check the structure of an art repository
}
