% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{repo_create}
\alias{repo_create}
\title{Create new art repository}
\usage{
repo_create(
  series,
  license = NULL,
  local_path = repo_local_path(),
  remote_url = repo_remote_url()
)
}
\arguments{
\item{series}{Name of the series (e.g., "series-rosemary")}

\item{license}{License type for the series ("ccby", "cc0", or "mit")}

\item{local_path}{Local folder in which to create the repository}

\item{remote_url}{Remote URL to check for a pre-existing repository, or NULL to skip check}
}
\value{
Invisibly returns TRUE on success, FALSE on failure (this may change)
}
\description{
Create new art repository
}
