% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\name{series_check}
\alias{series_check}
\title{Check the structure of an art series}
\usage{
series_check(series, origin = bucket_local_path())
}
\arguments{
\item{series}{Name of the series}

\item{origin}{Location in which to find the series}
}
\value{
Invisibly returns TRUE if all checks pass, FALSE if at least one
check fails
}
\description{
Check the structure of an art series
}
