% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{series_download}
\alias{series_download}
\title{Download an art series}
\usage{
series_download(series, dest_dir, base_url = NULL, keep_manifest = TRUE)
}
\arguments{
\item{series}{Name of the series}

\item{dest_dir}{Local folder into which the series folder is downloaded}

\item{base_url}{URL for the bucket containing all art series}

\item{keep_manifest}{Should the manifest file be downloaded? (default=TRUE)}
}
\value{
Tibble containing the download status data
}
\description{
Download an art series
}
