% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{status}
\alias{status}
\alias{bucket_status}
\alias{repo_status}
\title{Status of a repository or bucket}
\usage{
bucket_status(series, origin = bucket_local_path())

repo_status(series, origin = repo_local_path())
}
\arguments{
\item{series}{Name of the series}

\item{origin}{Location in which to find the bucket or repository folder}
}
\value{
Invisibly returns TRUE if no problems are detected, FALSE otherwise
}
\description{
Status of a repository or bucket
}
