% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{bucket_download}
\alias{bucket_download}
\title{Download series from remote bucket}
\usage{
bucket_download(
  series,
  destination = bucket_local_path(),
  origin = bucket_remote_path()
)
}
\arguments{
\item{series}{Name of the series}

\item{destination}{Local folder into which the series folder is downloaded}

\item{origin}{Location in which to find the series}
}
\value{
Tibble containing the download status data
}
\description{
Download series from remote bucket
}
