% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{manifest_read}
\alias{manifest_read}
\title{Read a series manifest file}
\usage{
manifest_read(series, origin = bucket_remote_path(), file = "manifest.csv")
}
\arguments{
\item{series}{Name of the series}

\item{origin}{Location in which to find the series}

\item{file}{File name for the manifest}
}
\value{
Tibble containing the manifest data
}
\description{
Read a series manifest file
}
