% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{bucket_remote_path}
\alias{bucket_remote_path}
\alias{bucket_local_path}
\alias{repo_remote_path}
\alias{repo_local_path}
\title{Paths to object buckets and repositories}
\usage{
bucket_remote_path(...)

bucket_local_path(...)

repo_remote_path(...)

repo_local_path(...)
}
\arguments{
\item{...}{Names of subfolders}
}
\value{
Fully qualified path or url
}
\description{
Paths to object buckets and repositories
}
