% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{series_download}
\alias{series_download}
\title{Download an art series}
\usage{
series_download(
  series,
  destination = bucket_local_path(),
  origin = bucket_remote_path()
)
}
\arguments{
\item{series}{Name of the series}

\item{destination}{Local folder into which the series folder is downloaded}

\item{origin}{Location in which to find the series}
}
\value{
Tibble containing the download status data
}
\description{
Download an art series
}
